+++ src/app/Smoothie/Conference/__tests__/Conference.test.js
