+++ src/app/Smoothie/Conference/__tests__/ListConference.test.js
