+++ src/app/Smoothie/__tests__/Smoothie.test.js
